Added scheduling for tasks and basic monitoring.

+++ src/net/nightwhistler/pageturner/library/CleanFilesTask.java
@@ -1,8 +1,9 @@
-package net.nightwhistler.pageturner.catalog;
+package net.nightwhistler.pageturner.library;
 
 import android.os.AsyncTask;
 import com.google.inject.Inject;
 import net.nightwhistler.pageturner.Configuration;
+import net.nightwhistler.pageturner.library.ImportCallback;
 import net.nightwhistler.pageturner.library.LibraryBook;
 import net.nightwhistler.pageturner.library.LibraryService;
 import net.nightwhistler.pageturner.library.QueryResult;
@@ -16,11 +17,19 @@ import java.util.List;
  */
 public class CleanFilesTask extends AsyncTask<Void, Void, Void> {
 
-    @Inject
     private Configuration config;
 
-    @Inject
     private LibraryService libraryService;
+
+    private ImportCallback callback;
+
+    private int deletedFiles = 0;
+
+    public CleanFilesTask(ImportCallback callback, LibraryService service, Configuration config) {
+        this.callback = callback;
+        this.libraryService = service;
+        this.config = config;
+    }
 
     @Override
     protected Void doInBackground(Void... voids) {
@@ -28,7 +37,7 @@ public class CleanFilesTask extends AsyncTask<Void, Void, Void> {
         QueryResult<LibraryBook> allBooks = libraryService.findAllByTitle(null);
         List<String> filesToDelete = new ArrayList<String>();
 
-        for ( int i=0; i < allBooks.getSize(); i++ ) {
+        for ( int i=0; i < allBooks.getSize() && ! isCancelled(); i++ ) {
             LibraryBook book = allBooks.getItemAt(i);
             File file = new File(book.getFileName());
 
@@ -40,9 +49,18 @@ public class CleanFilesTask extends AsyncTask<Void, Void, Void> {
         allBooks.close();
 
         for ( String fileName: filesToDelete ) {
-            libraryService.deleteBook(fileName);
+            if ( ! isCancelled() ) {
+                libraryService.deleteBook(fileName);
+                deletedFiles++;
+            }
         }
 
         return null;
+    }
+
+    @Override
+    protected void onPostExecute(Void aVoid) {
+        callback.taskCompleted(this, isCancelled());
+        callback.booksDeleted(deletedFiles);
     }
 }
